More minor adjustments
